Allow leading newlines inside arrays
The README.md has this at the end of the first TOML example:
```toml
hosts = [
  "alpha",
  "omega"
]
```
And later says:
>...arrays also ignore newlines....
But the ABNF doesn't allow for a newline to precede the first element.

I wasn't sure if the grammar should allow more than one newline preceding the first element, but given that an arbitrary number of whitespaces are allowed and the readme says that newlines are ignored in general I figured allowing an arbitrary number of newlines best matches your intent. It also won't break anyone that happens to have multiple leading newlines in their arrays.