Clarify that keys are always strings.
What
===
Add clarification that bare keys are always strings, even if they are
only digits.

Why
===
While it's implied that keys are all the same type and are of the string
type, it's not explicit in the specification. In some libraries using
the specification it's been suggested that bare keys could be
interpreted as other types, e.g. integers, but this is problematic as
most languages cannot effectively handle mixed type keys on maps. We
should be explicit that keys are only strings.