improve ABNF semantics on ml-basic-string escaped nl
this:
	string = """ escaped \

	     end"""

Was parsed as a single escaped newline, then as a newline part of the
string itself.

While it is fitting the role of rejecting invalid input and accepting
the valid one, it did not fit the role of parsing the content into
tokens.

To help the implementation, a multiline whitespace/newline token is
introduced: ml-basic-escaped-nl

This addresses issue #669.