Allow 0X, 0O, and 0B literals as capitals
Currently only "0x42" is accepted as a valid hex literal; "0X42" is
invalid.

I'm not sure if there's a good reason to not accept capital letters?
This is something that already works in many programming languages (not
all though, for example Go doesn't) and I don't see any downsides. I
can't find any previous discussions on this.

This also matches with the datetime parsing where "t" and "z" are
explicitly accepted as well (i.e. "1970-01-01t00:00:00z").

Since ABNF is case-insensitive we can use the "x" string for this
instead of specifying the codepoints, which is also a wee bit clearer.