Clarify that multi-line strings are not allowed in key names
There have been two cases recently where I had someone who was confused
about this.

This clarifies that multi-line strings are not allowed as key names for
the casual reader; the phrases "basic string" and "literal string" have
a specific meaning in the TOML spec, but that's not really obvious if
you're just looking up "what is the valid syntax for a key?" as the
rules are explained in a different section, and it requires some careful
reading.