Clarify tab can be escaped (or not) in basic strings
As discussed in #877, the current description of basic strings in the
docs say that tab characters are allowed. However this text is
immediately followed by an example that uses escaped tabs `\t`, which
make it a bit ambiguous and hard to understand.

The change introduced here explicitly says that tabs can be used either
directly or in the escaped form.