Fix use of "UTF-8"
UTF-8 is an encoding, not a character set, so "UTF-8 character" is not correct.

I picked the "should be implemented as" wording from floats, but am not sure this is what you want. It feels like there should be some consistent way to say "TOML interpreters should output $type as $format" for strings, integers, and floats (and maybe others...).