Expand changelog
One of my difficulties in updating TOML implementations in the past was
figuring out what exactly I need to change, as the changelog has
typically been extremely concise.

This expands things with commit references and some examples for changes
in syntax, which should make things a lot easier for people looking to
update TOML 1.0 implementations.